refactor(clawweb): consolidate shared runtime utilities
